API refactor module to make it stand alone
Previously the module relied on the existence of at least one service
module to carry out the work to generate the reports. This separation is a
little too much, so this commit merges `silverstripe-composer-versions`
into this one, creating a report instead of a model admin, which shows the
base information for the current install. It does this by destructuring
the composer.lock file, instead of reading the live composer information
via either the Packagist API or running composer via `exec`.