Add package summaries to the site summary report
For each individual module installed, list it's description as well as
it's title - which should also be listed in plain language, as opposed to
it's composer/packagist name (e.g. "Name" as opposed to
"vendor/silverstripe-name"). Filter the resulting report to only show
SilverStripe modules, identified by their "type" in packagist.