Repair CSV export formatting to be human friendly
As the work for formatting the on screen report saw the injection of HTML
into the summary, the exported CSV from the report saw the first field
return HTML (as expected) which isn't particularly meant for human
parsing.

We can fix this by defining fields for the report in the report, and
leaving the export fields (and other summaries) to give a less stylised
but ultimately more readable column format.