Add various extension points for security updates
In order to display information from
`bringyourownideas/silverstripe-composer-security-updates` we need to be
able to add various information to the summary, and badges which were
previously 'not a thing'. This commit makes badges 'a thing', makes them
extensible, and provides a rudimentary way to alter the summary
information for the on scren report.

Also added were namespaces for the GridFieldComponents, but this should
affect nothing (should not affect anything).