FIX pass correct parameter to badges extension hook
Although the function getBadges operated and returned correctly, the
extension point to hook the returned strucutre passed in the value of the
member property badges, then did nothing with the result. This resulted in
errors from extensions applied to the class expecting the parameter to be
an ArrayList they could push new badges into.